Refactor apartmentId type from Integer to Short in BookingDto and update MapperConfig converter accordingly

+++ src/main/java/com/verdemar/verdemar/domain/dto/BookingDto.java
@@ -4,7 +4,7 @@ import java.time.LocalDate;
 
 public class BookingDto {
     private Integer clientId;
-    private Integer apartmentId;
+    private Short apartmentId;
     private LocalDate startDate;
     private LocalDate endDate;
     private double totalPrice;
